requirements.txt
python-dotenv
pip
attrs
wheel
setuptools